Fix initial molecules positioning
When the application starts, the window and its contents are not
fully laid out yet, so  moleculeContainer size is initially zero
or very small. As a result, the initialization code cannot find
valid positions for all molecules within maxAttempts, leading to the warnings.

Solution: assume that intial window size is fixed.